Move invoice detail page from account to order.
