+++ ntemplates/account/invoices.html
@@ -2,7 +2,7 @@
 {% load static %}
 
 {% block content %}
-<div ng-controller="Invoices" ng-init="init({{ user.id }}, '{{ user.username }}')">
+<div ng-controller="UserInvoices" ng-init="init({{ user.id }}, '{{ user.username }}', '{% url 'order:show' '' %}')">
   <div class="page-header">
     <h3><i class="glyphicon glyphicon-time"></i> Your Invoices</h3>
   </div>
