+++ order/urls.py
@@ -4,7 +4,7 @@ urlpatterns = patterns('order.views',
   # Views
   url(r'^cart/$', 'view_cart', name='cart'),
   url(r'^checkout/$', 'checkout', name='checkout'),
-  url(r'^invoice/(?P<invoice_num>[\d\w-]+)/$', 'show_invoice', name='show'),
+  url(r'^invoice/(?P<id>\d*)$', 'show_invoice', name='show'),
 
   # PayPal
   url(r'^payment/paypal/execute/(?P<id>\d+)$', 'payment_paypal_execute', name='payment_paypal_execute'),
