+++ static/ng/account/controllers.js
@@ -1,18 +1,22 @@
 angular.module('account.controllers', [
   'common.resources',
-  'common.directive'
+  'account.directive'
 ])
-.controller('Invoices', [
+.controller('UserInvoices', [
   '$scope', '$q', 'InvoiceResource', 'OrderResource',
   function($scope, $q, InvoiceResource, OrderResource) {
     $scope.invoices = [];
+    $scope.invoiceUrlBase = null;
 
-    $scope.init = function(userId, username) {
+    $scope.init = function(userId, username, invoiceUrlBase) {
       var invoices;
 
       InvoiceResource.query({ user__username: username }).$promise
       .then(function(response) {
         invoices = response.results;
+        _.each(invoices, function(invoice) {
+          invoice.url = invoiceUrlBase + invoice.id;
+        });
         var orderPromises = _.flatten(_.map(invoices, function(invoice) {
           return _.map(invoice.orders, function(orderId, i) {
             var order = invoice.orders[i] = OrderResource.get({ id: orderId });
@@ -24,27 +28,6 @@ angular.module('account.controllers', [
       })
       .then(function() {
         $scope.invoices = invoices;
-      });
-    };
-  }
-])
-.controller('InvoiceDetail', [
-  '$scope', '$q', 'InvoiceResource', 'OrderResource',
-  function($scope, $q, InvoiceResource, OrderResource) {
-    $scope.invoice = null;
-
-    $scope.init = function(invoiceId) {
-      $scope.invoice = InvoiceResource.get({ id:invoiceId });
-      $scope.invoice.$promise
-      .then(function(invoice) {
-        if (_.size(invoice) < 1) {
-          // TODO: handle wrong response
-          return;
-        }
-
-        _.each(invoice.orders, function(orderId, i) {
-          invoice.orders[i] = OrderResource.get({ id: orderId });
-        });
       });
     };
   }
