+++ static/ng/order/index.js
@@ -1,0 +1,1 @@
+angular.module('order', ['order.controllers']);
